Rename `Endpoint` to `EndpointService`
